requirements.txt
dash>=3.0.0
pandas>=1.3.0
plotly>=5.9.0
python-dotenv>=1.0.0
